+++ CHANGELOG.md
@@ -26,6 +26,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Removed
 
 - Remove `fep-e232` cargo feature.
+- Drop support for `ethereum` OAuth flow.
 
 ### Fixed